add a simple multi-relation graph builder to relation dropdown menu
A new rlm_builder.js script is loaded once opening the relation dropdown menu.
jQuery retrieves data from the menu and store it in the browser's local storage.

In the dropdown menu, four new buttons and an input allow users to:
-“Add current individual” to a list (with an optional textual description).
-“Clear the graph” resets the constructed list of individuals.
-“Show the graph” prints a RLM graph between added individuals (if paths exist!)
-“Edit the graph” opens old m=TP&v=updRLM template to edit the made graph.
(Note that editing the graph this way does not modify initial storage.)

Note that a mix of indexes and individual keys is working and also that
unknowns individuals (? ?) always use indexes.